+++ src/DuneApp.cpp
@@ -976,7 +976,7 @@ void DuneApp::OnFileClose(MainWindow *window)
     delete window;
 #endif
     if (m_windows.size() == 0) {
-        for (int j = 0; j < m_filesToDelete.size(); j++)
+        for (size_t j = 0; j < m_filesToDelete.size(); j++)
             swRemoveFile(*m_filesToDelete[j]);
         swUploadCleanupPasswd(m_upload);
         Exit();
@@ -994,7 +994,7 @@ void DuneApp::OnFileExit()
     for (i = m_windows.first(); i; i = i->next()) {
         delete i->item();
     }
-    for (int j = 0; j < m_filesToDelete.size(); j++)
+    for (size_t j = 0; j < m_filesToDelete.size(); j++)
         swRemoveFile(*m_filesToDelete[j]);
     swUploadCleanupPasswd(m_upload);
     Exit();
@@ -1056,7 +1056,7 @@ void
 DuneApp::AddToFilesToDelete(char* string)
 {
     // test if string is already in list
-    for (int i = 0; i < m_filesToDelete.size(); i++)
+    for (size_t i = 0; i < m_filesToDelete.size(); i++)
        if (strcmp(string,(char*) m_filesToDelete[i])==0)
           return;
     m_filesToDelete.append(new MyString(string));
