+++ src/MFFloat.cpp
@@ -89,7 +89,7 @@ MFFloat::copy()
     if (m_value.size() == 0)
         return new MFFloat();
     float *value = new float[m_value.size()];
-    for (int i = 0; i < m_value.size(); i++)
+    for (size_t i = 0; i < m_value.size(); i++)
         value[i] = m_value[i];
     return new MFFloat(value, m_value.size());
 }
@@ -111,8 +111,8 @@ MFFloat::equals(const FieldValue *value) const
 bool
 MFFloat::equals(const MFFloat *value) const
 {
-    if (m_value.size() == value->getSize()) {
-        for (int i = 0; i < m_value.size(); i++)
+    if ((int)m_value.size() == value->getSize()) {
+        for (size_t i = 0; i < m_value.size(); i++)
             if (m_value[i] != value->getValue(i))
                 return false;
         return true;
@@ -144,13 +144,13 @@ void MFFloat::clamp(const FieldValue *min, const FieldValue *max)
 {
     if (min) {
         float fmin = ((SFFloat *) min)->getValue();
-        for (int i = 0; i < m_value.size(); i ++) {
+        for (size_t i = 0; i < m_value.size(); i ++) {
             if (m_value[i] < fmin) m_value[i] = fmin;
         }
     }
     if (max) {
         float fmax = ((SFFloat *) max)->getValue();
-        for (int i = 0; i < m_value.size(); i ++) {
+        for (size_t i = 0; i < m_value.size(); i ++) {
             if (m_value[i] > fmax) m_value[i] = fmax;
         }
     }
