+++ src/MFTime.cpp
@@ -68,7 +68,7 @@ FieldValue *
 MFTime::copy() 
 { 
     double *value = new double[m_value.size()];
-    for (int i = 0;i < m_value.size(); i++)
+    for (size_t i = 0; i < m_value.size(); i++)
         value[i] = m_value[i];
     return new MFTime(value, m_value.size());
 }
@@ -90,8 +90,8 @@ MFTime::equals(const FieldValue *value) const
 bool
 MFTime::equals(const MFTime *value) const
 {
-    if (m_value.size() == value->getSize()) {
-        for (int i = 0; i < m_value.size(); i++)
+    if ((int)m_value.size() == value->getSize()) {
+        for (size_t i = 0; i < m_value.size(); i++)
             if (m_value[i] != value->getValue(i))
                 return false;
         return true;
@@ -108,13 +108,13 @@ void MFTime::clamp(const FieldValue *min, const FieldValue *max)
 {
     if (min) {
         double dmin = ((SFTime *) min)->getValue();
-        for (int i = 0; i < m_value.size(); i++) {
+        for (size_t i = 0; i < m_value.size(); i++) {
             if (m_value[i] < dmin) m_value[i] = dmin;
         }
     }
     if (max) {
         double dmax = ((SFTime *) max)->getValue();
-        for (int i = 0; i < m_value.size(); i++) {
+        for (size_t i = 0; i < m_value.size(); i++) {
             if (m_value[i] > dmax) m_value[i] = dmax;
         }
     }
