+++ src/MeshMorphingNode.cpp
@@ -120,7 +120,7 @@ MeshMorphingNode::findInterpolators(InterpolatorInfo& info)
         }
     }
     
-    for (int i = 0; i < fieldList.size(); i++) {
+    for (size_t i = 0; i < fieldList.size(); i++) {
         Interpolator *inter = m_scene->findUpstreamInterpolator(nodeList[i], 
                                                                fieldList[i]);
         if (inter != NULL) {
@@ -154,7 +154,7 @@ MeshMorphingNode::addCoordinateInterpolator(Node *node,
     Node *interOutput = NULL;
     int interOutputEvent = -1;
     bool foundInterOutput = false;
-    for (int k = 0; k < info.interpolator.size(); k++) {
+    for (size_t k = 0; k < info.interpolator.size(); k++) {
         Node *node = info.interpolator[k];
         int eventIn = node->lookupEventIn("set_fraction", false);
         if (eventIn == INVALID_INDEX)
@@ -183,7 +183,7 @@ MeshMorphingNode::addCoordinateInterpolator(Node *node,
     do {
        foundMinKey = false;
        float minKey = 0;
-       for (int i = 0; i < info.interpolator.size(); i++)
+       for (size_t i = 0; i < info.interpolator.size(); i++)
            for (int j = 0; j < info.interpolator[i]->getNumKeys(); j++)
                if (info.interpolator[i]->getKey(j) > currentKey) {
                    if (foundMinKey == false) {
@@ -206,8 +206,8 @@ MeshMorphingNode::addCoordinateInterpolator(Node *node,
     void *data = initializeData();
     if (data == NULL)
         return true;
-    for (int i = 0; i < keys.size(); i++) {
-        for (int j = 0; j < info.interpolator.size(); j++) {
+    for (size_t i = 0; i < keys.size(); i++) {
+        for (size_t j = 0; j < info.interpolator.size(); j++) {
             Interpolator *inter = info.interpolator[j];
             int field = info.field[j];
             if (inter != NULL)
