+++ src/NodeGeoLocation.cpp
@@ -117,7 +117,7 @@ NodeGeoLocation::preDraw()
     glPushMatrix();
     transform();
 
-    for (int i = 0; i < childList->size(); i++)
+    for (size_t i = 0; i < childList->size(); i++)
         if (childList->get(i) != this)
             childList->get(i)->preDraw();
