+++ src/NodeGeoTransform.cpp
@@ -142,7 +142,7 @@ NodeGeoTransform::preDraw()
     glPushMatrix();
     transform();
 
-    for (int i = 0; i < childList->size(); i++)
+    for (size_t i = 0; i < childList->size(); i++)
         if (childList->get(i) != this)
             childList->get(i)->preDraw();
